Bugfix 1. prevent hashing password from uploading video, only when change password and creating account 2. preotect accessing delete/edit url from users who is not the owner

+++ src/controllers/videoController.js
@@ -12,8 +12,6 @@ export const watch = async(req, res) => {
     const {id} = req.params;
     // find video which has a specific id from video object and populate("owner")
     const video = await Video.findById(id).populate("owner");
-    
-    console.log(video);
 
     if(!video) {
         return res.render("404", {pageTitle: "Video Not Found"})
@@ -24,23 +22,35 @@ export const watch = async(req, res) => {
 
 export const getEdit = async(req, res) => {
     const {id} = req.params;
+    const {user : {_id}} = req.session;
     // access DB : get selected id video object 
     const video= await Video.findById(id);
-
+    
     if(!video) {
         return res.status(404).render("404", {pageTitle: "Video Not Found"})
     }
+
+    // protect to access edit page if user is not owner of the video
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+
     return res.render("edit", {pageTitle: `Editing ${video.title}`, video})
 }
 
 export const postEdit = async(req, res) => {
     const {id} = req.params;
+    const {user : {_id}} = req.session;
     // get edited input value 
     const {title, description, hashtags} = req.body;
     // find Video by id if video exists
     const video = await Video.exists({ _id: id });
     if(!video) {
         return res.status(404).render("404", {pageTitle: "Video Not Found"})
+    }
+    // protect to access postEdit page if user is not owner of the video
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     // update new value in DB
     await Video.findByIdAndUpdate(id,{
@@ -102,6 +112,22 @@ export const search = async(req, res) => {
 
 export const deleteVideo = async(req, res) => {
     const {id} = req.params;
+    const {user : {_id}} = req.session;
+
+    const video = await Video.findById(id);
+    
+    //if video doesnt exist -> 404 
+    if(!video) {
+        return res.status(404).render("404", {pageTitle: "Video Not Found"})
+    }
+
+    // if user is not owner of video redirect to home 
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+    
+    // delete video 
     await Video.findByIdAndDelete(id);
+    
     return res.redirect("/");
 }
